+++ js/data.js
@@ -1,0 +1,18 @@
+const TYPE = ["palace", "flat", "house", "bungalow", "hotel"];
+const CHECKIN = ["12:00", "13:00", "14:00"];
+const CHECKOUT = ["12:00", "13:00", "14:00"];
+const FEATURES = [
+  "wifi",
+  "dishwasher",
+  "parking",
+  "washer",
+  "elevator",
+  "conditioner",
+];
+const PHOTOS = [
+  "https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg",
+  "https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg",
+  "https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg",
+];
+
+export {TYPE, CHECKIN, CHECKOUT, FEATURES, PHOTOS}
