+++ js/getRandomArrayLength.js
@@ -1,0 +1,6 @@
+import {getRange} from './getRandomNumb.js';
+
+const getRandomArrayLength = (elements) =>
+  elements.slice(0, getRange(0, elements.length));
+
+export {getRandomArrayLength}
